Update code commented, restructure imports, and refactor classes
The commit introduces detailed code comments for better understanding of the JavaScript functions in index.html. Moreover, the imports in 'server/main.py' are reorganized based on their categories. Changes have also been made to the classes in 'utils/history.py'. Lastly, 'langchain-openai' was added to the setup.py dependencies.